+++ start/templatetags/custom_tags.py
@@ -1,3 +1,5 @@
+import math
+
 from django import template
 from django.template import engines
 from django.utils.html import format_html
@@ -6,9 +8,25 @@ from django.utils.safestring import mark_safe
 register = template.Library()
 
 
+@register.filter
+def get_flag(obj, attr):
+    return obj[attr]
+
+
+@register.simple_tag
+def tuple_args(dictionary, a, b):
+    if (a, b) in dictionary.keys():
+        return math.sqrt(dictionary[(a, b)]/1000000000)
+    else:
+        return 0
+
+
 @register.filter()
 def get_value(dictionary, key):
-    return int(10 * float(dictionary.get(key)))
+    if key in dictionary.keys():
+        return int(10 * float(dictionary.get(key)))
+    else:
+        return 0
 
 
 @register.filter()
@@ -62,7 +80,10 @@ def get_item(dictionary, key):
 
 @register.filter()
 def get_list_item(List, i):
-    return List[int(i)]
+    if i < len(List):
+        return List[int(i)]
+    else:
+        return List[0]
 
 
 @register.filter()